% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impor.R
\name{impor}
\alias{impor}
\title{Impor dataset}
\source{
Dataset bersumber dari Open Data Kota Jakarta \url{http://data.jakarta.go.id}.
}
\usage{
impor(.data)
}
\arguments{
\item{.data}{data keluaran dari fungsi \code{cari()} (data dengan kelas jktry)}
}
\value{
dataframe atau list berisi dataframe apabila dataset yang hendak diimpor jumlahnya lebih dari satu.
}
\description{
Fungsi untuk mengimpor dataset hasil pencarian dari fungsi \code{cari()}.
}
\examples{

library(jakartaraya)

kebakaran <- cari(kata_kunci = "kebakaran")
kebakaran

dataset_kebakaran <- impor(kebakaran)
dataset_kebakaran

impor(kebakaran[1,])

}
