% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jakartaraya-package.R
\docType{package}
\name{jakartaraya}
\alias{jakartaraya}
\alias{jakartaraya-package}
\title{paket \code{bandungjuara}}
\description{
Antarmuka untuk Open Data Kota Jakarta
}
\details{
Silakan baca dokumentasi di
\href{https://github.com/aswansyahputra/jakartaraya#readme}{GitHub}
}
\author{
\strong{Maintainer}: Muhammad Aswan Syahputra \email{aswansyahputra@sensolution.id}

}
